Edit margin and placement in footer
Changed vertical margin of contacts and address in footer to 0. This can
be done as the buttons have enough margin to still push down end of page
enough. This was done as there was too much space in between elements
when the footer was viewed on a mobile vertical screen. Also Changed
justify-content:flex-start to space-around on buttons in footer to
provide more balanced placement on both mobile and pc.